third_party/WebKit/LayoutTests/external/wpt/webstorage/storage_string_conversion-expected.txt

This is a testharness.js-based test.
FAIL localStorage only stores stringsassert_equals: expected "function(){}" but got "function (){}"
FAIL sessionStorage only stores stringsassert_equals: expected "function(){}" but got "function (){}"
Harness: the test ran to completion.
